% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seasonality.R
\name{add_HeadTail}
\alias{add_HeadTail}
\title{Add one year data in the head and tail}
\usage{
add_HeadTail(d, nptperyear = 23)
}
\arguments{
\item{d}{A data.table, should have \code{t} (compositing date) column 
(\code{Date} variable).}
}
\value{
data.table
}
\description{
Add one year data in the head and tail
}
