% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_class_deprecated.R
\name{bisection}
\alias{bisection}
\title{bisection}
\usage{
bisection(f, a, b, tolx = 1e-07, toly = 0.001, maxit = 1000, trace = T,
  ln = TRUE, ...)
}
\description{
bisection
}
\examples{
bisection(f_whitdf, 10, 1e6, toly =1e-1)
}
\references{
[1]. https://rpubs.com/aaronsc32/bisection-method-r
}
