% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_MOD13A1.R
\docType{data}
\name{CA_NS6}
\alias{CA_NS6}
\title{MOD13A1 EVI observations at flux site CA-NS6}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 161 rows and 6 columns.
}
\usage{
data('CA_NS6')
}
\description{
Variables in CA-NS6:
\itemize{
\item \code{site}: site name
\item \code{y}: EVI
\item \code{date}: date of image
\item \code{t}: date of compositing image
\item \code{w}: weights of data point
\item \code{QC_flag}: QC flag of y, in the range of \code{c("snow", "cloud", "shadow", "aerosol", "marginal", "good")}
}
}
\keyword{datasets}
