% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_remote_functions.R
\name{check_dir}
\alias{check_dir}
\title{check_dir}
\usage{
check_dir(indir)
}
\arguments{
\item{indir}{which directory to detect in remote computer}
}
\value{
if indir directory if exist finally (TRUE or FALSE)
}
\description{
check indir directory if exist in remote computer, if not create it
}
