% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_fit}
\alias{check_fit}
\title{check_fit}
\usage{
check_fit(yfit, ylu)
}
\arguments{
\item{yfit}{Numeric vector, curve fitting result}

\item{ylu}{limits of y value, [ymin, ymax]}
}
\description{
Curve fitting values are constrained in the range of \code{ylu}.
Only constrain trough value for a stable background value. But not for peak
value.
}
