% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_tools.R
\name{check_GS_HeadTail}
\alias{check_GS_HeadTail}
\title{Check growing season head and tail minimum values}
\usage{
check_GS_HeadTail(pos, ypred, minlen, A = NULL, ...)
}
\arguments{
\item{pos}{data.frame, with the columns of \code{pos}, \code{type}, and \code{val}.}

\item{minlen}{\code{nptperyear/3}, distance from peak point.}
}
\value{
di
}
\description{
Check growing season head and tail minimum values
}
\keyword{internal}
