% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics_fitting.R
\name{check_input}
\alias{check_input}
\title{check_input}
\usage{
check_input(t, y, w, missval, maxgap = 10, trim = TRUE, alpha = 0.01)
}
\arguments{
\item{t}{A numeric vector}

\item{y}{A numeric vector}

\item{w}{A numeric vector}

\item{missval}{Double, which is used to replace NA values in y. If missing,
the default vlaue is `ylu[1] - diff(ylu)/10`.}

\item{maxgap}{Integer, nptperyear/4 will be a suitable value. If continuous
missing value numbers less than maxgap, then interpolate those NA values by
zoo::na.approx; If false, then replace those NA values with a const value,
`ylu[1]`}
}
\description{
If just replace NA values with a const missing value, it is inappropriate for
middle growing season points. If interpolating all values by na.approx, it is
unsuitable for large number continous missing segments, e.g. in the start
or end of growing season. So a maxgap parameter with a default value of
`nptperyear/4` will be a good choice.
}
