% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_season}
\alias{check_season}
\title{check_season}
\usage{
check_season(d, rm_closed = TRUE, rtrough_max = 0.7, r_min = 0.02)
}
\arguments{
\item{d}{Data.frame of growing season dividing info}

\item{rtrough_max}{\code{ytrough <= rtrough_max*A}, A is the amplitude of y.}

\item{r_min}{Threshold is defined as the difference of peak value with
trough value. There are two threshold (left and right). The minimum threshold
should be greater than r_min.}
}
\description{
check_season
}
\keyword{internal}
