% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seasonality.R
\name{check_seasonality}
\alias{check_seasonality}
\title{Check vegetation seasonality}
\usage{
check_seasonality(INPUT, IsPlot = F, plotdat = INPUT, nf = 3, ...)
}
\arguments{
\item{INPUT}{A list object with the elements of 't', 'y', 'w', 'Tn' (option)
and 'ylu', returned by \code{check_input}.}

\item{IsPlot}{Boolean}

\item{plotdat}{A list or data.table, with 't', 'y' and 'w'. Only if
IsPlot=true, plotdata will be used to plot. Known that y and w in \code{INPUT}
have been changed, we suggest using the original data.table.}

\item{nf}{the parameter of \code{wHANTS}, number of frequencies to be
considered above the zero frequency}

\item{...}{For 'season_3y', Other parameters passed to `season`;
For `season`, other parameters passed to findpeaks.}
}
\description{
Check vegetation seasonality
}
