% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_ylu}
\alias{check_ylu}
\title{check_ylu}
\usage{
check_ylu(yfit, ylu)
}
\arguments{
\item{yfit}{Numeric vector, curve fitting result}

\item{ylu}{limits of y value, [ymin, ymax]}
}
\value{
yfit, the numeric vector in the range of \code{ylu}.
}
\description{
Curve fitting values are constrained in the range of \code{ylu}.
Only constrain trough value for a stable background value. But not for peak
value.
}
\examples{
check_ylu(1:10, c(2, 8))
}
