% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_remote_functions.R
\name{cluster_Init}
\alias{cluster_Init}
\title{Initial Clusters}
\usage{
cluster_Init(ncluster = 4, pkgs, vars, expr)
}
\description{
Initial clusters for parallel computing based on parallel package
}
\examples{
cluster_Init(pkgs = "httr",
     expr = source("R/mainfunc/main.R", encoding = "utf-8"))
}
