% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative.R
\name{curvature}
\alias{curvature}
\title{curvature}
\usage{
curvature(fit, analytical = TRUE, smoothed.spline = FALSE, ...)
}
\arguments{
\item{fit}{A curve fitting object returned by \code{curvefit}.}

\item{analytical}{If true, \code{numDeriv} package \code{grad} and \code{hess}
will be used; if false, \code{D1} and \code{D2} will be used.}

\item{smoothed.spline}{Whether apply \code{smooth.spline} first?}

\item{...}{Other parameters will be ignored.}
}
\description{
curvature
}
