% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvefits_extend.R
\name{curvefit}
\alias{curvefit}
\title{curve fit vegetation index (VI) time-series}
\usage{
curvefit(y, t = index(y), tout = t, meth = "BFGS",
  methods = c("AG", "Beck", "Elmore", "Gu", "Klos", "Zhang"), ...)
}
\arguments{
\item{y}{Vegetation time-series index, numeric vector}

\item{t}{The corresponding doy of x}

\item{tout}{The output interpolated time.}

\item{meth}{optimization method}

\item{methods}{Fine curve fitting methods, can be one or more of \code{c('AG', 
'Beck', 'Elmore', 'Gu', 'Klos', 'Zhang')}.}

\item{...}{other parameters passed to curve fitting function.}
}
\value{
fits
}
\description{
curve fit VI using  methods
}
\note{
'Klos' and 'Gu' have many parameters. It will be slow and not stable.
}
\seealso{
\code{\link{FitAG}}, \code{\link{FitDL.Beck}}, 
\code{\link{FitDL.Elmore}}, \code{\link{FitDL.Gu}}, 
\code{\link{FitDL.Klos}}, \code{\link{FitDL.Zhang}}
}
