% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvefit.R
\name{curvefit0}
\alias{curvefit0}
\title{curvefit0}
\usage{
curvefit0(
  y,
  t = index(y),
  tout = t,
  methods = c("AG", "Beck", "Elmore", "Gu", "Klos", "Zhang"),
  w = NULL,
  ...
)
}
\description{
curvefit0
}
\keyword{internal}
