% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvefits.R
\name{curvefit_optimx}
\alias{curvefit_optimx}
\title{curvefit_optimx}
\usage{
curvefit_optimx(x, t = index(x), tout = t, optimFUN = I_optimx,
  methods = c("spline", "beck", "elmore", "AG"), meth, pfun, ...)
}
\arguments{
\item{optimFUN}{`I_optim` or `I_optimx`}

\item{meth}{c('BFGS','CG','Nelder-Mead','L-BFGS-B','nlm','nlminb',
spg','ucminf','Rcgmin','Rvmmin','newuoa','bobyqa','nmkb','hjkb')}

\item{pfun}{c(p_nlminb, p_ncminf, p_nlm, p_optim)}
}
\description{
With the help of `optimx` package, try to find which optimization function
is best.
}
