% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract_main.R
\name{curvefit_site}
\alias{curvefit_site}
\title{curvefit_site}
\usage{
curvefit_site(t, y, w, nptperyear = 46, wFUN = wTSM, iters = 2, lambda,
  south = FALSE, IsPlot = FALSE, methods = c("AG", "zhang", "beck",
  "elmore", "Gu"), debug = FALSE, ...)
}
\arguments{
\item{t}{A date vector}

\item{y}{A numberic vector, same length as t}

\item{methods}{A character vector, c('spline', 'beck', 'elmore', 'klos', 'AG', 'zhang')}

\item{...}{other parameters pass to season or curvefit_site}
}
\description{
curvefit_site
}
