% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvefits.R
\name{curvefits2}
\alias{curvefits2}
\title{curvefits2}
\usage{
curvefits2(t, y, w, nptperyear = 46, wFUN = wTSM, iters = 2, lambda,
  south = FALSE, IsPlot = FALSE, Aymin_less = 0.6, ymax_min = 0.1,
  methods = c("AG", "zhang", "beck", "elmore", "Gu"), debug = FALSE, ...)
}
\arguments{
\item{t}{A date vector}

\item{y}{A numberic vector, same length as t}

\item{methods}{A character vector, c('spline', 'beck', 'elmore', 'klos', 'AG', 'zhang')}

\item{...}{other parameters pass to season or curvefit_site}
}
\description{
curvefits2
}
