% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvefits_LocalModel.R
\name{cutoff}
\alias{cutoff}
\title{cutoff}
\usage{
cutoff(n1, n2, k = n1:n2)
}
\arguments{
\item{n1, n2}{peak and trough (or trough and peak)}
}
\value{
A numeric vector, in small intervals around \code{(tL + tC)/2} and \code{(tC+tR)/2},
respectively, smoothly drop from 1 to 0.
}
\description{
cutoff
}
\references{
\enumerate{
\item Per J\"onsson, P., Eklundh, L., 2004. TIMESAT - A program for analyzing
time-series of satellite sensor data. Comput. Geosci. 30, 833-845.
https://doi.org/10.1016/j.cageo.2004.05.006.
}
}
\keyword{internal}
