% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{cv_coef}
\alias{cv_coef}
\title{weighted CV}
\usage{
cv_coef(x, w)
}
\arguments{
\item{x}{Numeric vector}

\item{w}{weights of different point}
}
\value{
Named numeric vector, (mean, sd, cv).
}
\description{
weighted CV
}
\examples{
library(phenofit)
x = rnorm(100)
coefs <- cv_coef(x)
}
\keyword{internal}
