% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit.R
\name{D1.phenofit}
\alias{D1.phenofit}
\title{D1.phenofit}
\usage{
\method{D1}{phenofit}(fit, numDeriv = FALSE, smspline = FALSE, ...)
}
\arguments{
\item{smspline}{If smspline = TRUE or attr(fit$fun, 'gradient') == null,
it will use smooth.spline to get der1 and der2.}
}
\description{
D1.phenofit
}
