% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit.R
\name{grad.phenofit}
\alias{grad.phenofit}
\alias{hess.phenofit}
\alias{D1.phenofit}
\alias{D2.phenofit}
\alias{curvature.phenofit}
\title{Gradient (grad) and Hessian (hess) based on \code{numDeriv} package}
\usage{
\method{grad}{phenofit}(fit, tout)

hess.phenofit(fit, tout)

\method{D1}{phenofit}(fit, numDeriv = FALSE, smspline = FALSE, ...)

\method{D2}{phenofit}(fit, numDeriv = FALSE, smspline = FALSE, ...)

\method{curvature}{phenofit}(fit, numDeriv = FALSE, smspline = FALSE,
  ...)
}
\arguments{
\item{fit}{Curve fitting result returned by \code{curvefit}.}

\item{tout}{A vector of time steps at which the function can be predicted}

\item{numDeriv}{If true, \code{numDeriv} package \code{grad} and \code{hess} 
will be used; if false, \code{D1} and \code{D2} will be used.}

\item{smspline}{If smspline = TRUE or attr(fit$fun, 'gradient') == null,
it will use smooth.spline to get der1 and der2.}

\item{...}{Other parameters are ignored.}
}
\description{
Gradient (grad) and Hessian (hess) based on \code{numDeriv} package
}
