% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_seasons.R
\name{divide_seasons}
\alias{divide_seasons}
\title{divide_seasons}
\usage{
divide_seasons(
  d_obs,
  nptperyear = 23,
  iters = 2,
  south = FALSE,
  lambda = 100,
  rFUN = "smooth_wWHIT",
  wFUN = wTSM,
  calendarYear = FALSE,
  maxExtendMonth = 12,
  .v_curve = FALSE,
  is.plot = FALSE,
  ...
)
}
\arguments{
\item{.v_curve}{If true, it will use V-curve theory to optimize Whittaker
parameter, lambda.}

\item{d}{data.frame, with the columns of \code{t}, \code{y} and \code{w}.}

\item{sp}{one line data.frame}

\item{sitename}{character}

\item{.movmean}{moving mean by \code{wSG}}
}
\description{
divide_seasons
}
\note{
site-year may be not continuous.
}
