% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics.R
\name{doubleLog.zhang}
\alias{doubleLog.zhang}
\title{simplest piecewise logistics function}
\usage{
doubleLog.zhang(par, t)
}
\description{
fit whole growing season NDVI. First introduced to phenology by
 Zhang Xiaoyang, RME, 2003
}
