% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_main.R
\name{ExtractPheno}
\alias{ExtractPheno}
\title{ExtractPheno}
\usage{
ExtractPheno(fits, TRS = c(0.1, 0.2, 0.5, 0.6), IsPlot = FALSE)
}
\arguments{
\item{outdir}{save calculated phenology in oudir txts}
}
\description{
Get yearly vegetation phenological metrics of a curve fitting method
}
