% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics.R
\name{f_goal}
\alias{f_goal}
\title{Common goal function of those curve fitting methods}
\usage{
f_goal(par, y, t, fun = c(doubleLog.Elmore, doubleLog.Beck,
  doubleLog.Klos, doubleLog.Gu, doubleLog.Zhang, doubleAG), w, ylu, ...)
}
\arguments{
\item{par}{A vector of parameters}

\item{y}{Numeric vector, vegetation index time-series}

\item{t}{Numeric vector, \code{Date} variable}

\item{fun}{A curve fitting function, see \code{\link{Logistic}}.}

\item{w}{(optional) Numeric vector, weights of \code{y}. If not specified, 
weights of all \code{NA} values will be \code{wmin}, the others will be 1.0.}

\item{ylu}{\code{ymin, ymax}, which is used to force \code{ypred} in the 
range of \code{ylu}.}

\item{...}{other parameters will be ignored.}
}
\value{
RMSE root mean square error of curve fitting values.
}
\description{
Common goal function of those curve fitting methods
}
