% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics_cpp.R
\name{f_goal2}
\alias{f_goal2}
\title{objective function of double logistics}
\usage{
f_goal2(par, fun, y, t, ypred, w = NULL, ylu = NULL, ...)
}
\arguments{
\item{par}{A vector of parameters}

\item{fun}{A curve fitting function, can be one of \code{doubleAG},
\code{doubleLog.Beck}, \code{doubleLog.Elmore}, \code{doubleLog.Gu},
\code{doubleLog.Klos}, \code{doubleLog.Zhang}, see \code{\link[=Logistic]{Logistic()}}
for details.}

\item{y}{Numeric vector, vegetation index time-series}

\item{t}{Numeric vector, \code{Date} variable}

\item{w}{(optional) Numeric vector, weights of \code{y}. If not specified,
weights of all \code{NA} values will be \code{wmin}, the others will be 1.0.}

\item{ylu}{\code{ymin, ymax}, which is used to force \code{ypred} in the
range of \code{ylu}.}

\item{...}{others will be ignored.}
}
\description{
objective function of double logistics
}
\keyword{internal}
