% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_fFIT.R
\name{fFIT}
\alias{fFIT}
\title{S3 class of fine curve fitting object.}
\format{\describe{
  \item{tout}{Corresponding doy of prediction}
  \item{zs}{curve fitting values of every iteration}
  \item{ws}{weight of every iteration}
  \item{par}{Optimized parameter of fine curve fitting method}
  \item{fun}{The name of fine curve fitting function.}
}}
\description{
\code{fFIT} is returned by \code{\link{optim_pheno}}.
}
