% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_input.R
\name{find_season.peaks}
\alias{find_season.peaks}
\alias{find_season.default}
\title{find_season}
\usage{
find_season.peaks(rfit, info_peak, options = list(), ...)

find_season.default(
  ypred,
  t = seq_along(ypred),
  nptperyear = NULL,
  south = NULL,
  options = list(),
  ...
)
}
\arguments{
\item{rfit}{data.frame with the columns of t and \code{ziter...}, the first column
should be \code{t}, and the last should be \code{ziter...}.}
}
\description{
find_season
}
\keyword{internal}
