% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoBrks.R
\name{findpeaks}
\alias{findpeaks}
\title{findpeaks}
\usage{
findpeaks(x, nups = 1, ndowns = nups, zero = "0", peakpat = NULL,
  minpeakheight = -Inf, minpeakdistance = 1, threshold_min = 0,
  threshold_max = 0, npeaks = 0, sortstr = FALSE)
}
\description{
findpeaks
}
