% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics_fitting.R
\name{FitDL.Beck}
\alias{FitDL.Beck}
\title{Fitting double logistics, asymmetric gaussian functions}
\usage{
FitDL.Beck(x, t = index(x), tout = t, optimFUN = I_optimx,
  method = "nlminb", w, ...)
}
\arguments{
\item{x}{input vegetation index time-series.}

\item{t}{the corresponding doy(day of year) of x.}

\item{tout}{the output curve fitting time-series time steps.}

\item{optimFUN}{optimization function to solve curve fitting functions'
parameters. It's should be `optimx_fun`, or `optim_p`.}

\item{method}{method passed to `optimx` or `optim` function.}

\item{...}{other paraters passed to optimFUN, such as weights.}
}
\value{
list(pred, par, fun)
}
\description{
Fitting double logistics, asymmetric gaussian functions
}
\examples{
FitDL.Beck  (x, t, tout, optimFUN = optim_p, pfun = p_nlminb)
FitDL.Elmore(x, t, tout, optimFUN = optim_p, pfun = p_nlminb)
FitDL.Gu    (x, t, tout, optimFUN = optim_p, pfun = p_nlminb)
FitDL.Klos  (x, t, tout, optimFUN = optim_p, pfun = p_optim, method = 'BFGS')

FitDL.Zhang (x, t, tout, optimFUN = optim_p, pfun = p_nlm)
FitAG (x, t, tout, optimFUN = optim_p, pfun = p_nlminb)
}
\references{
[1]. Beck, P.S.A., Atzberger, C., Hogda, K.A., Johansen, B., Skidmore, A.K., 
     2006. Improved monitoring of vegetation dynamics at very high latitudes: 
     A new method using MODIS NDVI. Remote Sens. Environ. 
     https://doi.org/10.1016/j.rse.2005.10.021
}
