% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_process.R
\name{get_date_AVHRR}
\alias{get_date_AVHRR}
\title{get_date_AVHRR}
\usage{
get_date_AVHRR(year_begin = 1982, year_end = 2015)
}
\arguments{
\item{year_begin}{integer}

\item{year_end}{integer}
}
\description{
Generate image dates from \code{year_begin} to \code{year_end}.
This function is only for AVHRR satellites.
}
\examples{
date_AVHRR <- get_date_AVHRR()
}
