% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_remote_functions.R
\name{get_files}
\alias{get_files}
\title{get_files}
\usage{
get_files(indir, pattern = "*.txt", full.names = F)
}
\arguments{
\item{indir}{directory where stored results}

\item{pattern}{certain file with assign pattern returned by dir function in remote computer indir}
}
\description{
store calculated result in remote computers, and retrieve them later
}
