% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_sbatch.R
\name{get_slurm_out}
\alias{get_slurm_out}
\title{get_slurm_out
Merge the slurm result.}
\usage{
get_slurm_out(indir = ".", pattern = "result.*.RDS",
  outfile = "result.rda", IsSave = TRUE)
}
\description{
get_slurm_out
Merge the slurm result.
}
