% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_QC.R
\name{getBits}
\alias{getBits}
\title{Extract bitcoded QA information from bin value}
\usage{
getBits(x, start, end = start)
}
\arguments{
\item{x}{Binary value}

\item{start}{Bit starting position}

\item{end}{Bit ending position}
}
\value{
decimal value
}
\description{
Extract bitcoded QA information from bin value
}
