% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract_main.R
\name{getFits_pheno}
\alias{getFits_pheno}
\title{getFits_pheno}
\usage{
getFits_pheno(fits, TRS = c(0.1, 0.2, 0.5, 0.6), IsPlot = FALSE)
}
\arguments{
\item{outdir}{save calculated phenology in oudir txts}
}
\description{
Get yearly vegetation phenological metrics of a curve fitting method
}
