% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenofit.R
\name{getFitValueYears}
\alias{getFitValueYears}
\title{getFitValueYears}
\usage{
getFitValueYears(fit_years)
}
\description{
getFitValueYears
}
\examples{
origin.date <- fit$INPUT$t[1]
out$t \%<>\% add(origin.date - 1)
}
