% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvefit.R
\name{getparam}
\alias{getparam}
\alias{getparams}
\title{Get parameters from curve fitting result}
\usage{
getparam(fit)

getparams(fits)
}
\arguments{
\item{fit}{Curve fitting result by \code{curvefits} result.}

\item{fits}{Multiple methods curve fitting results by \code{curvefits} result.}
}
\description{
Get parameters from curve fitting result
}
