% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_main.R
\name{getparams}
\alias{getparams}
\title{Get parameters from multiple curve fitting results}
\usage{
getparams(fits)
}
\arguments{
\item{fits}{List Object of curvefits returned}
}
\description{
Get parameters from multiple curve fitting results
}
