% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_MODIS.R
\name{getRealDate}
\alias{getRealDate}
\title{getRealDate}
\usage{
getRealDate(df)
}
\arguments{
\item{df}{A data.table, at least with columns of \code{date}, and
\code{DayOfYear}.}
}
\value{
A data.table with a new column \code{t}, which is the exact compositing date.
}
\description{
convert MODIS \code{DayOfYear} to the exact compositing date.
}
