% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract_main.R
\name{getYears_pheno}
\alias{getYears_pheno}
\title{getYears_pheno}
\usage{
getYears_pheno(lst, tt, nptperyear, nyear, years, outdir = "./",
  IsPlot = FALSE, IsSave = FALSE)
}
\arguments{
\item{outdir}{save calculated phenology in oudir txts}
}
\description{
Deprecated function.
}
\details{
Get multi-years vegetation phenology metrics
}
