% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{GOF}
\alias{GOF}
\title{GOF}
\usage{
GOF(Y_obs, Y_sim, w, include.cv = FALSE)
}
\arguments{
\item{Y_obs}{Numeric vector, observations}

\item{Y_sim}{Numeric vector, corresponding simulated values}

\item{w}{Numeric vector, weights of every points}

\item{include.cv}{If true, cv will be returned.}
}
\value{
\itemize{
\item \code{RMSE} root mean square error
\item \code{NSE} NASH coefficient
\item \code{R2} correlation of determination
\item \code{Bias} bias
\item \code{Bias_perc} bias percentage 
\item \code{MAE} mean absolute error 
\item \code{pvalue} pvalue of \code{R}
\item \code{n_sim} number of valid obs.
\item \code{R} pearson correlation
}
}
\description{
Good of fitting
}
