% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{grad_par}
\alias{grad_par}
\title{grad_par}
\usage{
grad_par(FUN)
}
\arguments{
\item{par}{Named vector, or named list. It can't be matrix.
Colnamed matrix cooperating with plyr::alply will be fine.}
}
\value{
return the gradient function of double logistics
}
\description{
grad_par
}
