% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{gradf_t}
\alias{gradf_t}
\title{DERIVS from the aspect of par}
\usage{
gradf_t(FUN)
}
\description{
For par aspect DERIVS, it had better to use deriv return function.
For t aspect DERIVS, it had better to use D, then eval D expression
}
