% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_functions.R
\name{gridSaveToPoly}
\alias{gridSaveToPoly}
\title{gridSaveToPoly}
\usage{
gridSaveToPoly(grid, file)
}
\arguments{
\item{grid}{The spatialPixelDataframe or spatialGridDataframe object.}

\item{file}{The string of output shapefile name including path.}
}
\description{
Write spatialPixelDataframe into polygon shpfile for the convenience of
checking in google earth.
}
