% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_HANTS.R
\name{HANTS2}
\alias{HANTS2}
\title{Weighted HANTS SMOOTH}
\usage{
HANTS(nptperyear,nf,y,ts,HiLo,low,high,fet,dod,delta)
}
\arguments{
\item{y}{array of input sample values (e.g. NDVI values)}

\item{ts}{array of size ni of time sample indicators}

\item{nf}{number of frequencies to be considered above the zero frequency}

\item{ylu}{[low, high] of time-series y (values outside the valid range are rejeced}

\item{nptperyear}{length of the base period, measured in virtual samples}

\item{ni}{nr. of images (total number of actual samples of the time}
}
\value{
amp returned array of amplitudes, first element is the average of
        the curve
phi returned array of phases, first element is zero
yr array holding reconstructed time series
}
\description{
Modified by Dongdong Kong, 2018-05-29
}
\author{
Wout Verhoef, NLR, Remote Sensing Dept. June 1998
Mohammad Abouali (2011), Converted to MATLAB
Dongdong Kong (2018), introduced to R and modified into weighted model.
}
