% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimDL.R
\name{I_optim}
\alias{I_optim}
\title{I_optim}
\usage{
I_optim(prior, FUN, y, t, tout, pfun = p_optim, method = "BFGS", ...)
}
\arguments{
\item{pfun}{with prefixion of 'p_', pfun was self modified and unified
optimization function for double logistics and many other functions.}

\item{method}{was only used for `p_optim` function. Other pfun, e.g. p_nlm, p_nlminb
will ignore it.}
}
\description{
Interface of self defined optimization functions.
}
