% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_whit_lambda.R
\name{init_lambda}
\alias{init_lambda}
\title{Initial lambda value of whittaker taker}
\usage{
init_lambda(y)
}
\arguments{
\item{y}{Numeric vector}
}
\description{
Initial lambda value of whittaker taker
}
