% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics-init_param.R
\name{init_param}
\alias{init_param}
\alias{init_Zhang}
\alias{init_AG}
\alias{init_AG2}
\alias{init_Beck}
\alias{init_Elmore}
\alias{init_Gu}
\alias{init_Klos}
\title{init_param}
\usage{
init_param(y, t, w)

init_Zhang(y, t = index(y), w)

init_AG(y, t = index(y), w)

init_AG2(y, t = index(y), w)

init_Beck(y, t = index(y), w)

init_Elmore(y, t = index(y), w)

init_Gu(y, t = index(y), w)

init_Klos(y, t = index(y), w)
}
\arguments{
\item{y}{Numeric vector, vegetation index time-series}

\item{t}{Numeric vector, \code{Date} variable}

\item{w}{(optional) Numeric vector, weights of \code{y}. If not specified,
weights of all \code{NA} values will be \code{wmin}, the others will be 1.0.}
}
\description{
Initialize parameters of double logistic function
}
\examples{
library(phenofit)
# simulate vegetation time-series
fFUN = doubleLog.Beck
par  = c(
    mn  = 0.1,
    mx  = 0.7,
    sos = 50,
    rsp = 0.1,
    eos = 250,
    rau = 0.1)
t    <- seq(1, 365, 8)
tout <- seq(1, 365, 1)
y <- fFUN(par, t)

l_param <- init_param(y, t)
}
\keyword{internal}
