% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_data.R
\docType{data}
\name{input_single}
\alias{input_single}
\title{input object with one growing season per year}
\format{
An object of class \code{list} of length 6.
}
\usage{
data('input_single')
}
\description{
Variables in \code{input_single}:
\itemize{
\item \code{t}: date of compositing image
\item \code{y}: EVI
\item \code{w}: weights of data point
\item \code{ylu}: lower and upper boundary
\item \code{nptperyear}: points per year
\item \code{south}: boolean, whether in south Hemisphere?
}
}
\keyword{datasets}
