% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{kurtosis}
\alias{kurtosis}
\alias{skewness}
\title{skewness and kurtosis}
\usage{
kurtosis(x, na.rm = FALSE, type = 3)

skewness(x, na.rm = FALSE, type = 3)
}
\arguments{
\item{x}{a numeric vector containing the values whose skewness is to be
computed.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}

\item{type}{an integer between 1 and 3 selecting one of the algorithms for
computing skewness.}
}
\description{
Inherit from package \code{e1071}
}
\examples{
x = rnorm(100)
coef_kurtosis <- kurtosis(x)
coef_skewness <- skewness(x)

}
\keyword{internal}
