% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{listk}
\alias{listk}
\alias{list.cbind}
\alias{list.rbind}
\title{listk}
\usage{
listk(...)

list.cbind(x)

list.rbind(x)
}
\arguments{
\item{...}{objects, possibly named.}

\item{x}{A list object, with data.frame of element}
}
\description{
listk
}
