% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract.R
\name{localMinMax}
\alias{localMinMax}
\title{Local maxima and minima based on Tommy's solution. And only kept the best if
two points was so close.}
\usage{
localMinMax(x, meth = c("min", "max"), MinSep = 15)
}
\arguments{
\item{x}{numeric vector}

\item{meth}{A string parameter, only can be 'min' or 'max'.}

\item{MinSep}{The Minimum gap of two local minimum (maximum).}
}
\description{
maxval: f' = 0, f'_left > 0 & f'_right < 0
minval: f' = 0, f'_left < 0 & f'_right > 0
}
\details{
https://stackoverflow.com/questions/6836409/finding-local-maxima-and-minima
If having NA values means phenology result was nquestionable.
}
