% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics.R
\name{Log.zhang}
\alias{Log.zhang}
\title{Define double logistics, piecewise logistics and many other functions used to
curve fit VI time-series}
\usage{
doubleLog.beck(par, t)
doubleLog.gu(par, t)
doubleLog.elmore(par, t)
doubleLog.klos(par, t)
doubleLog.zhang(par, t) #piecewise
doubleAG(par, t)        #piecewise

All of those function have `par` and `formula` attributes for the convenience for
analytical D1 and D2
only fit part of growing season NDVI, before or after peak NDVI
}
\description{
Define double logistics, piecewise logistics and many other functions used to
curve fit VI time-series
}
