% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics.R
\name{Logistic}
\alias{Logistic}
\alias{doubleLog.zhang}
\alias{doubleAG}
\alias{doubleLog.beck}
\alias{doubleLog.elmore}
\alias{doubleLog.gu}
\alias{doubleLog.klos}
\title{Double logistics functions}
\usage{
Logistic(par, t)

doubleLog.zhang(par, t)

doubleAG(par, t)

doubleLog.beck(par, t)

doubleLog.elmore(par, t)

doubleLog.gu(par, t)

doubleLog.klos(par, t)
}
\description{
Define double logistics, piecewise logistics and many other functions to
curve fit VI time-series
\itemize{
   \item \code{Logistic} The traditional simplest logistic function. It can 
     be only used in half growing season, i.e. vegetation green-up or senescence 
     period. 
   \item \code{doubleLog.zhang} Piecewise logistics, Zhang Xiaoyang, RME, 2003.
   \item \code{doubleAG} Asymmetric Gaussian.
   \item \code{doubleLog.beck} Beck logistics.
   \item \code{doubleLog.gu} Gu logistics.
   \item \code{doubleLog.elmore} Elmore logistics.
   \item \code{doubleLog.klos} Klos logistics.
}
}
\details{
All of those function have \code{par} and \code{formula} attributes for the 
convenience for analytical D1 and D2
}
\references{
Peter M. Atkinson, et al., 2012, RSE, 123:400-417
}
