% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{merge_pdf}
\alias{merge_pdf}
\title{merge_pdf}
\usage{
merge_pdf(outfile = "RPlot.pdf", indir = "Figure", pattern = "*.pdf",
  del = FALSE)
}
\description{
rely on python pdfmerge package, `pip install pdfmerge`
}
