% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{MissInfo}
\alias{MissInfo}
\alias{MissInfo-package}
\alias{MissInfo-package}
\title{MissInfo}
\description{
detect missing information of meteorological data and interpolate missing values using common interpolation methods
}
\keyword{information,}
\keyword{interpolate}
\keyword{missing}
