% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_MOD13A1.R
\docType{data}
\name{MOD13A1}
\alias{MOD13A1}
\title{MOD13A1}
\format{
An object of class \code{list} of length 2.
}
\usage{
data('MOD13A1')
}
\description{
A data.table dataset, raw data of MOD13A1 data, clipped in 10 representative points
('DE-Obe', 'IT-Col', 'CN-Cha', 'AT-Neu', 'ZA-Kru', 'AU-How', 'CA-NS6',
'US-KS2', 'CH-Oe2', 'CZ-wet').
}
\details{
Variables in MOD13A1:
\itemize{
\item \strong{\code{dt}}: vegetation index data
\itemize{
\item \code{system:index}: image index
\item \code{DayOfYear}: Numeric, Julian day of year
\item \code{DayOfYear}: corresponding doy of compositing NDVI and EVI
\item \code{DetailedQA}: VI quality indicators
\item \code{SummaryQA}: Quality reliability of VI pixel
\item \code{EVI}: Enhanced Vegetation Index
\item \code{NDVI}: Normalized Difference Vegetation Index
\item \code{date}: Date, corresponding date
\item \code{site}: String, site name
\item \code{sur_refl_b01}: Red surface reflectance
\item \code{sur_refl_b02}: NIR surface reflectance
\item \code{sur_refl_b03}: Blue surface reflectance
\item \code{sur_refl_b07}: MIR surface reflectance
\item \code{.geo}: geometry
}
\item \strong{\code{st}}: station info
\itemize{
\item \code{ID}: site ID
\item \code{site}: site name
\item \code{lat}: latitude
\item \code{lon}: longitude
\item \code{IGBPname}: IGBP land cover type
}
}
}
\references{
\enumerate{
\item https://code.earthengine.google.com/dataset/MODIS/006/MOD13A1
}
}
\keyword{datasets}
