% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{movmean}
\alias{movmean}
\title{movmean}
\usage{
movmean(y, halfwin = 1L, SG_style = FALSE, w = NULL)
}
\arguments{
\item{y}{A numeric vector.}

\item{halfwin}{Integer, half of moving window size}

\item{SG_style}{If true, head and tail values will be in the style of SG
(more weights on the center point), else traditional moving mean style.}

\item{w}{Corresponding weights of yy, same long as yy.}
}
\description{
NA and Inf values in the yy will be ignored automatically.
}
\examples{
x <- 1:100
x[50] <- NA; x[80] <- Inf
s1 <- movmean(x, 2, SG_style = TRUE)
s2 <- movmean(x, 2, SG_style = FALSE)
}
