% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{obj.size}
\alias{obj.size}
\title{obj.size}
\usage{
obj.size(obj, unit = "Mb")
}
\arguments{
\item{obj}{Object}

\item{unit}{"Kb", "Mb" or "Gb"}
}
\description{
Get object size in \code{unit}
}
\examples{
obj.size(1:100)
}
