% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_find.R
\name{opt_season}
\alias{opt_season}
\title{Growing season division (unstable version)}
\usage{
opt_season(INPUT, options = NULL, verbose = FALSE, ...)
}
\arguments{
\item{INPUT}{A list object with the elements of \code{t}, \code{y}, \code{w},
\code{Tn} (optional) and \code{ylu}, returned by \code{\link[=check_input]{check_input()}}.}

\item{options}{see details}

\item{...}{others to \code{\link[=season]{season()}}}
}
\description{
Growing season division (unstable version)
}
\keyword{internal}
