% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimDL.R
\name{p_nlminb}
\alias{p_nlminb}
\alias{p_ncminf}
\alias{p_nlm}
\alias{p_optim}
\title{Unified optim function}
\usage{
p_nlminb(par0, objective, ...)

p_ncminf(par0, objective, ...)

p_nlm(par0, objective, ...)

p_optim(par0, objective, method = "BFGS", ...)
}
\arguments{
\item{par0}{Initial values for the parameters to be optimized over.}

\item{objective}{A function to be minimized (or maximized), with first 
argument the vector of parameters over which minimization is to take place. 
It should return a scalar result.}

\item{...}{other parameters passed to \code{objective}.}

\item{method}{optimization method to be used in \code{p_optim}. See 
\code{\link[stats]{optim}}.}
}
\value{
\describe{
   \item{convcode}{ An integer code. 0 indicates successful convergence. 
   Various methods may or may not return sufficient information to allow all 
   the codes to be specified. An incomplete list of codes includes }
   \item{value}{ The value of fn corresponding to par }
   \item{par}{ The best parameter found }
   \item{nitns}{ the number of iterations }
   \item{fevals}{ The number of calls to fn. }
   \item{gevals}{ The number of calls to gr. This excludes those calls needed
   to compute the Hessian, if requested, and any calls to fn to compute a 
   finite-difference approximation to the gradient. }
}
}
\description{
\itemize{
   \item \code{p_nlminb} Optimization using PORT routines, see 
                         \code{\link[stats]{nlminb}}.
   \item \code{p_ncminf} General-Purpose Unconstrained Non-Linear 
                         Optimization, see \code{\link[ucminf]{ucminf}}.
   \item \code{p_nlm} Non-Linear Minimization, \code{\link[stats]{nlm}}.
   \item \code{p_optim} General-purpose Optimization, see 
                        \code{\link[stats]{optim}}.
}
}
