% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_sbatch.R
\name{par_sbatch}
\alias{par_sbatch}
\title{par_sbatch}
\usage{
par_sbatch(X, FUN, ..., nodes = 8, cpus_per_node = 16)
}
\arguments{
\item{X}{a vector (atomic or list) or an expressions vector. Other objects
(including classed objects) will be coerced by as.list.}

\item{FUN}{the function to be applied to (mclapply) each element of X or
(mcmapply) in parallel to ....}

\item{...}{other parameters passed to mclapply}

\item{nodes}{The (maximum) number of cluster nodes to spread the calculation
over. slurm_apply automatically divides params in chunks of approximately
equal size to send to each node. Less nodes are allocated if the parameter
set is too small to use all CPUs on the requested nodes.}

\item{cpus_per_node}{The number of CPUs per node on the cluster;
determines how many processes are run in parallel per node.}
}
\description{
R parallel function, designed for linux sbatch model.
}
