% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract.R
\name{PhenoTrs}
\alias{PhenoTrs}
\alias{PhenoDeriv}
\alias{PhenoGu}
\alias{PhenoKl}
\title{Phenology extraction functions}
\usage{
PhenoTrs(fit, approach = c("White", "Trs"), trs = 0.5, IsPlot = TRUE, ...)

PhenoDeriv(fit, IsPlot = TRUE, smspline = TRUE, show.lgd = T, ...)

PhenoGu(fit, IsPlot = TRUE, smspline = TRUE, ...)

PhenoKl(fit, IsPlot = TRUE, show.lgd = T, ...)
}
\arguments{
\item{fit}{Curve fitting result returned by \code{curvefit}.}

\item{...}{Other parameters are ignored.}

\item{smspline}{If smspline = TRUE or attr(fit$fun, 'gradient') == null,
it will use smooth.spline to get der1 and der2.}
}
\description{
Phenology extraction functions

PhenoDeriv
}
