% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{phenofit}
\alias{phenofit}
\alias{phenofit-package}
\alias{phenofit-package}
\title{phenofit}
\description{
Vegetation phenology package
}
\keyword{Vegetation}
\keyword{package}
\keyword{phenology}
