% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_loaddata.R
\name{phenofit_loaddata}
\alias{phenofit_loaddata}
\title{update all INPUT data according to \code{input} file.}
\usage{
phenofit_loaddata(options, rv, ...)
}
\arguments{
\item{options}{should has children of \code{file_site}, and one of
\code{file_veg_rda} or \code{file_veg_text}.}

\item{rv}{return values to reactiveValues object.}

\item{...}{ignored.}
}
\description{
update all INPUT data according to \code{input} file.
}
\keyword{internal}
