% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_process.R
\name{phenofit_plot}
\alias{phenofit_plot}
\title{phenofit_plot}
\usage{
phenofit_plot(obj, type = "all", methods)
}
\arguments{
\item{obj}{\code{fFIT}}

\item{type}{one of c("season", "fitting", "pheno", "all")}
}
\description{
phenofit_plot
}
