% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_process.R
\name{phenofit_process}
\alias{phenofit_process}
\alias{getsite_data}
\alias{getsite_INPUT}
\alias{phenofit_finefit}
\title{phenofit_process}
\usage{
getsite_data(df, sitename, dateRange = NULL)

getsite_INPUT(df, st, sitename, nptperyear, dateRange = NULL)

phenofit_finefit(INPUT, brks, options, TRS = c(0.2, 0.5, 0.6), ...)

phenofit_process(options, dateRange = c(as.Date("2010-01-01"),
  as.Date("2014-12-31")), nsite = -1, .progress = NULL,
  .parallel = FALSE, ...)
}
\arguments{
\item{df}{data.table of vegetation time-series. At least with the columns of
\code{t, y, w}.}

\item{sitename}{character}

\item{dateRange}{Date vector, \code{[date_begin, date_end]}. filter input in the
range of \code{dateRange}}

\item{st}{data.table of site information, e.g. \code{site}, \code{lat}.}

\item{nptperyear}{Integer, number of images per year.}

\item{INPUT}{An object returned by \code{check_input}}

\item{brks}{object returned by \code{\link[=season_mov]{season_mov()}} and \code{\link[=season]{season()}}}

\item{options}{options of phenofit needed to export.}

\item{TRS}{Threshold for \code{PhenoTrs}.}

\item{...}{ignored}

\item{nsite}{the max number of sites to process. \code{-1} means all sites.}

\item{.progress}{boolean}

\item{.parallel}{boolean}
}
\description{
phenofit_process

phenofit_process
}
\examples{
\dontrun{
file_json <- system.file('shiny/phenofit/perference/phenofit_setting.json', package = "phenofit")
options <- setting.read(file_json)
r <- phenofit_process(options, nsite=2)
}
}
