% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_process.R
\name{phenofit_season}
\alias{phenofit_season}
\title{phenofit_season}
\usage{
phenofit_season(INPUT, options, IsPlot = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{INPUT}{An object returned by \code{check_input}}

\item{options}{options of phenofit}

\item{IsPlot}{whether to plot season dividing procedure?}

\item{verbose}{boolean. Whether print parameters of \code{season_mov} or \code{season}?}

\item{...}{other parameters to \code{\link[=season_mov]{season_mov()}} or \code{\link[=season]{season()}}}
}
\value{
same object returned by \code{\link[=season_mov]{season_mov()}} and \code{\link[=season]{season()}}
}
\description{
\itemize{
\item \code{phenofit_season}: Calculate growing season dividing information
\item \code{phenofit_finefit}: Fine fitting
}
}
