% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_shiny.R
\name{phenofit_shiny}
\alias{phenofit_shiny}
\title{phenofit shiny app}
\usage{
phenofit_shiny()
}
\description{
The GUI allows you to interactively visualize curve fitting time series and 
phenological metrics.
}
\examples{
\dontrun{
phenofit_shiny
}
}
\keyword{GUI,}
\keyword{end,}
\keyword{front}
\keyword{interactive}
