% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_process.R
\name{phenofit_TS.avhrr}
\alias{phenofit_TS.avhrr}
\title{phenofit_TSM.avhrr}
\usage{
phenofit_TS.avhrr(options, dateRange = c(as.Date("2010-01-01"),
  as.Date("2014-12-31")), nsite = -1, outdir = ".",
  exportType = "all", .progress = NULL, .parallel = FALSE, ...)
}
\arguments{
\item{exportType}{could be one of \code{"all", "pheno"}. If "all" used, all 
also will be exported. Note that exported fitting series is daily scale, which 
is quite large.}
}
\description{
phenofit_TSM.avhrr
}
