% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract.R
\name{PhenoPlot}
\alias{PhenoPlot}
\title{PhenoPlot}
\usage{
PhenoPlot(t, y, main = "", ...)
}
\arguments{
\item{t}{Numeric vector, \code{Date} variable}

\item{y}{Numeric vector, vegetation index time-series}

\item{main}{figure title}

\item{...}{ignored parameters}
}
\description{
PhenoPlot
}
