% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit.R
\name{plot.phenofit}
\alias{plot.phenofit}
\title{plot function for phenofit class}
\usage{
\method{plot}{phenofit}(x, ...)
}
\arguments{
\item{...}{Other parameters are ignored.}
}
\description{
plot curve fitting VI, gradient (first order difference D1), hessian (D2),
curvature (k) and the change rate of curvature(der.k)
}
