% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_input.R
\name{plot_input}
\alias{plot_input}
\title{Plot INPUT returned by check_input}
\usage{
plot_input(INPUT, wmin = 0.2, ...)
}
\arguments{
\item{INPUT}{A list object with the elements of \code{t}, \code{y}, \code{w}, 
\code{Tn} (optional) and \code{ylu}, returned by \code{\link{check_input}}.}

\item{wmin}{Double, minimum weigth (i.e. weight of snow, ice and cloud).}

\item{...}{other parameter will be ignored.}
}
\description{
Plot INPUT returned by check_input
}
