% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenofit.R
\name{plot_phenofit}
\alias{plot_phenofit}
\title{plot_phenofit}
\usage{
plot_phenofit(df_fit, seasons, title = NULL, font.size = 14,
  show.legend = TRUE)
}
\arguments{
\item{df_fit}{data.frame of curve fittings returned by \code{\link{get_fitting}}.}

\item{seasons}{Growing season dividing object returned by \code{\link{season}}
and \code{\link{season_3y}}.}

\item{title}{String, title of figure.}

\item{font.size}{Font size of axis.text}

\item{show.legend}{Boolean}
}
\description{
plot_phenofit
}
