% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_season.R
\name{plot_season}
\alias{plot_season}
\title{plot_season}
\usage{
plot_season(INPUT, brks, plotdat, ylu, IsPlot.OnlyBad = FALSE,
  show.legend = TRUE)
}
\arguments{
\item{brks}{A list object returned by \code{season_mov}.}

\item{ylu}{limits of y value, [ymin, ymax]}

\item{show.legend}{Whether to show legend?}
}
\description{
Plot growing season divding result.
}
