% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_input.R
\name{plotdata}
\alias{plotdata}
\title{Plot INPUT returned by check_input}
\usage{
plotdata(INPUT, wmin = 0.2, ...)
}
\arguments{
\item{INPUT}{A list object with the elements of 't', 'y', 'w', 'Tn' (option)
and 'ylu', returned by \code{check_input}.}

\item{wmin}{Double, minimum weigth (i.e. weight of snow, ice and cloud).}

\item{...}{other parameter will be ignored.}
}
\description{
Plot INPUT returned by check_input
}
