% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit.R
\name{print.phenofit}
\alias{print.phenofit}
\title{Display phenofit object}
\usage{
print.phenofit(x, ...)
}
\arguments{
\item{x}{Curve fitting object returned by \code{curvefit}.}

\item{...}{ignored.}
}
\description{
Display phenofit object
}
