% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_functions.R
\docType{data}
\name{prj84}
\alias{prj84}
\title{prj84
Default Spatial object projection: WGS84.}
\format{An object of class \code{CRS} of length 1.}
\usage{
prj84
}
\description{
prj84
Default Spatial object projection: WGS84.
}
\keyword{datasets}
