% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_vars.R
\name{qc_levels}
\alias{qc_levels}
\alias{qc_colors}
\alias{qc_shapes}
\title{qc level, color and shape}
\usage{
qc_levels

qc_colors

qc_shapes
}
\description{
qc level, color and shape
}
\keyword{internal}
