% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{R2_sign}
\alias{R2_sign}
\title{Determinated correlation critical value}
\usage{
R2_sign(n, NumberOfPredictor = 2, alpha = 0.05)
}
\arguments{
\item{n}{length of observation.}

\item{NumberOfPredictor}{Number of predictor.}

\item{alpha}{significant level.}
}
\value{
F statistic and R2 at significant level
}
\description{
Determinated correlation critical value
}
