% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/season_mov.R
\name{rcpp_season_filter}
\alias{rcpp_season_filter}
\alias{check_season_dt}
\alias{check_season_list}
\title{season_filter}
\usage{
rcpp_season_filter(d, rm_closed = TRUE, rtrough_max = 0.7, r_min = 0.02)

check_season_dt(dt)

check_season_list(lst_dt)
}
\arguments{
\item{d}{Data.frame of growing season dividing info}

\item{rtrough_max}{\code{ytrough <= rtrough_max*A}, A is the amplitude of y.}

\item{r_min}{Threshold is defined as the difference of peak value with
trough value. There are two threshold (left and right). The minimum threshold
should be greater than r_min.}

\item{dt}{data.table of growing season division info}

\item{lst_dt}{list of \code{dt}. Every year is corresponding to a \code{dt}.}
}
\description{
season_filter
}
\keyword{internal}
