% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_wSG}
\alias{rcpp_wSG}
\alias{rcpp_SG}
\title{Weighted Savitzky-Golay written in RcppArmadillo}
\usage{
rcpp_wSG(y, halfwin = 1L, d = 1L, w = NULL)

rcpp_SG(y, halfwin = 1L, d = 1L)
}
\arguments{
\item{y}{colvec}

\item{halfwin}{halfwin of Savitzky-Golay}

\item{d}{polynomial of degree. When d = 1, it becomes moving average.}

\item{w}{colvec of weight}
}
\description{
NA and Inf values in the yy has been ignored automatically.
}
\examples{
y <- 1:15
w <- seq_along(y)/length(y)

frame = 5
d = 2
s1 <- rcpp_wSG(y, frame, d, w)
s2 <- rcpp_SG(y, frame, d)
}
