% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_functions.R
\name{rdist.earth}
\alias{rdist.earth}
\title{rdist.earth}
\usage{
rdist.earth(x1, x2 = NULL)
}
\arguments{
\item{x1}{A position matrix [lat, lon], degree unit.}

\item{x2}{A position matrix [lat, lon], degree unit.}
}
\description{
rdist.earth
}
