% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%<-\%}
\alias{\%>\%}
\alias{\%<>\%}
\alias{make_date}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lubridate}{\code{\link[lubridate:make_date]{make_date()}}}

  \item{magrittr}{\code{\link[magrittr:\%<>\%]{\%<>\%}}, \code{\link[magrittr:\%>\%]{\%>\%}}}

  \item{zeallot}{\code{\link[zeallot:\%<-\%]{\%<-\%}}}
}}

