% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_tools.R
\name{removeClosedExtreme}
\alias{removeClosedExtreme}
\title{rm too closed peaks or troughs}
\usage{
removeClosedExtreme(pos, ypred, A = NULL, y_min)
}
\arguments{
\item{pos}{data.table with the columns of \verb{"val", "pos", "left", "right", "type"}.}
}
\value{
A data.table with the columns of:
\verb{"val", "pos", "left", "right", "type"}
}
\description{
rm too closed peaks or troughs
}
\examples{
#removeClosedExtreme(pos, ypred, A, y_min)

}
\keyword{internal}
