% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seasonality.R
\name{rmNonSeasonality}
\alias{rmNonSeasonality}
\title{REMOVE non-seasonality points according to HANTS smooth result}
\usage{
rmNonSeasonality(dt, IsPlot = T, file = "SI.pdf")
}
\description{
REMOVE non-seasonality points according to HANTS smooth result
}
