% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoBrks.R
\name{season}
\alias{season}
\title{season}
\usage{
season(INPUT, lambda, nptperyear = 46, south = FALSE, iters = 2,
  wFUN = wTSM, IsPlot = TRUE, minpeakdistance = nptperyear/6,
  ymax_min = 0.5, Aymin_less = 0.6, TRS = 0.05, meth = c("whit", "sg"),
  ..., max_MaxPeaksperyear = 2, max_MinPeaksperyear = 3)
}
\arguments{
\item{INPUT}{returned list(t, y, w, ylu) object from `check_input`}

\item{TRS}{multiply A (diff(ylu)) becomes the threshold parameter of
`findpeak` function}

\item{...}{Other parameters passed to findpeaks, e.g. 'threshold',
'minpeakheight', 'minpeakdistance'}
}
\value{
list(dt, di)
$whit
}
\description{
First smooth VI timeseries by weighted whittaker, then use findpeak function
to get the local maximum and local minimum values. Two local minimum defined
a growing season. If two local minimum(maximum) was adjacent, then only the
smaller(biger) is left.
}
\details{
Then according to season pos, based to local maximum position divide yearly
growing season. lambda need to set carefully.

divide growing season circle just like TIMESAT season function
}
