% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_input.R
\name{season_input}
\alias{season_input}
\title{Growing season division (unstable version)}
\usage{
season_input(INPUT, options = NULL, verbose = FALSE, ...)
}
\arguments{
\item{INPUT}{A list object with the elements of \code{t}, \code{y}, \code{w},
\code{Tn} (optional) and \code{ylu}, returned by \code{\link[=check_input]{check_input()}}.}

\item{options}{see the following section \verb{options for season} for details.}

\item{...}{others parameter to \code{\link[=set_options]{set_options()}}}
}
\description{
Growing season division (unstable version)
}
\keyword{internal}
