% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_remote_functions.R
\name{selectHost}
\alias{selectHost}
\title{selectHost}
\usage{
selectHost(cl)
}
\arguments{
\item{cl}{returned by makePSOCKcluster}
}
\value{
matched unique hosts Id positions
}
\description{
select unique hosts to excute initial check functions
}
