% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_setting.R
\name{setting}
\alias{setting}
\alias{setting.get}
\alias{setting.read}
\alias{setting.write}
\title{setting}
\usage{
setting.get(input)

setting.read(file)

setting.write(pars, file)
}
\description{
setting

get parameters of phenofit shinyapp
}
\examples{
\dontrun{
pars = get_setting(input)
print(str(pars))
write_json(setting, "phenofit_setting.json", pretty = TRUE)
}
}
