% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_sgfitw.R
\name{sgfitw}
\alias{sgfitw}
\title{Weighted Savitzky-Golay}
\usage{
sgfitw(y, w, nptperyear, ylu, wFUN = wTSM, iters = 2,
  frame = floor(nptperyear/7) * 2 + 1, d = 2, ...)
}
\description{
Weighted Savitzky-Golay
}
