% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_class_deprecated.R
\name{smooth_HdH}
\alias{smooth_HdH}
\title{Faster whittake smoother and yhat}
\usage{
smooth_HdH(y, w = 0 * y + 1, lambda = 10000)
}
\description{
Faster whittake smoother and yhat
}
\references{
[1]. L- and V-curves for optimal smoothing, Statistical Modelling 2015; 15(1): 91-111
}
