% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smooth_wSG}
\alias{smooth_wSG}
\alias{smooth_SG}
\title{Weighted Savitzky-Golay}
\usage{
smooth_wSG(y, w, halfwin, d = 2L)

smooth_SG(y, halfwin, d = 2L)
}
\arguments{
\item{y}{colvec}

\item{w}{colvec of weight}

\item{halfwin}{halfwin of Savitzky-Golay}

\item{d}{polynomial of degree. When d = 0, it becomes moving average.}
}
\description{
Weighted Savitzky-Golay
}
\examples{
y <- 1:15
w <- seq_along(y)/length(y)

frame = 5
d = 2
s1 <- smooth_wSG(y, w, frame, d)
s2 <- smooth_SG(y, frame, d)
}
