% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics_fitting.R
\name{splinefit}
\alias{splinefit}
\title{using cubic spline function to avoid the difficult in setting parameter
lambda in smooth.spline}
\usage{
splinefit(y, t = index(y), tout = t, plot = FALSE,
  df.factor = 0.06, ...)
}
\description{
cubic spline is inappropriate for daily inputs. Its smooth is not enough.
On the contrary, smooth.spline with a low freedom can smooth well.
}
