% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit.R
\name{statistic.phenofit}
\alias{statistic.phenofit}
\title{statistic}
\usage{
statistic.phenofit(fit)
}
\arguments{
\item{fit}{The phenofit object}
}
\description{
statistics to evaluate curve fitting performance
}
