% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenofit.R
\name{tidyFitPheno}
\alias{tidyFitPheno}
\title{tidyFits_pheno}
\usage{
tidyFitPheno(pheno, origin)
}
\arguments{
\item{pheno}{Extracted phenology from \code{ExtractPheno}}

\item{origin}{time origin, i.e. \code{INPUT$t[1]}}
}
\description{
Tidy for every method with multiple years phenology data
}
