% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{reorder_name}
\alias{reorder_name}
\alias{rm_empty}
\alias{contain}
\title{reorder_name}
\usage{
reorder_name(d, headvars = c("site", "date", "year", "doy", "d8", "d16"),
  tailvars = "")

rm_empty(x)

contain(d, pattern = "NDVI|EVI")
}
\arguments{
\item{d}{A data.frame vector, or list}

\item{headvars}{headvars will be in the head columns.}

\item{tailvars}{tailvars will be in the tail columns.}

\item{x}{A vector or list}

\item{pattern}{string used to match \code{names(d)}}
}
\description{
reorder_name

rm_empty

contain
find assigned pattern variable names
}
