% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightUpdate.R
\name{wBisquare}
\alias{wBisquare}
\title{Bisquare weight update method}
\usage{
wBisquare(y, yfit, w, ...)
}
\description{
Bisquare has been modified to emphasis on upper envelope.
}
\references{
[1]. https://au.mathworks.com/help/curvefit/smoothing-data.html#bq_6ys3-3 \cr
[2]. Garcia, D., 2010. Robust smoothing of gridded data in one and higher
dimensions with missing values. Computational statistics & data analysis,
54(4), pp.1167-1178.
}
