% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightUpdate.R
\name{wChen}
\alias{wChen}
\title{Chen et al., (2004) weights updating method}
\usage{
wChen(y, yfit, w, ...)
}
\description{
Chen et al., (2004) weights updating method
}
\references{
[1]. Chen, J., J\"onsson, P., Tamura, M., Gu, Z., Matsushita, B., Eklundh, L., 
     2004. A simple method for reconstructing a high-quality NDVI time-series 
     data set based on the Savitzky-Golay filter. Remote Sens. Environ. 91, 
     332–344. https://doi.org/10.1016/j.rse.2004.03.014
}
