% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_class.R
\name{whitsmw2}
\alias{whitsmw2}
\title{Weigthed Whittaker Smoother}
\usage{
whitsmw2(y, w, ylu, nptperyear, wFUN = wTSM, iters = 1, lambdas = 1000,
  df = NULL, ...)
}
\description{
Weigthed Whittaker Smoother
}
\references{
[1]. Eilers, P.H.C., 2003. A perfect smoother. Anal. Chem. https://doi.org/10.1021/ac034173t
[2]. Frasso, G., Eilers, P.H.C., 2015. L- and V-curves for optimal smoothing. Stat. 
     Modelling 15, 91–111. https://doi.org/10.1177/1471082X14549288
}
