% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whit_class_deprecated.R
\name{whitsmw2_d}
\alias{whitsmw2_d}
\title{whittaker smoother}
\usage{
whitsmw2_d(y, w, ylu, nptperyear, wFUN = wTSM, iters = 1, lambdas = 1000,
  df = NULL, ..., validation = FALSE)
}
\description{
whittaker smoother
}
