% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightUpdate.R
\name{wTSM}
\alias{wTSM}
\title{wTSM}
\usage{
wTSM(y, yfit, w, iter = 2, nptperyear, wfact = 0.5, ...)
}
\description{
Weight updating method in TIMESAT
}
\references{
[1]. Per J\"onsson, P., Eklundh, L., 2004. TIMESAT - A program for analyzing
    time-series of satellite sensor data. Comput. Geosci. 30, 833-845.
    https://doi.org/10.1016/j.cageo.2004.05.006
}
\author{
Per J\"onsson, Malm\"o University, Sweden \email{per.jonsson@ts.mah.se} \cr
    Lars Eklundh, Lund University, Sweden \email{lars.eklundh@nateko.lu.se} \cr
    Translate by Dongdong Kong, 01 May 2018.
}
