% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seasonality.R
\name{add_HeadTail}
\alias{add_HeadTail}
\title{Add one year data in the head and tail}
\usage{
add_HeadTail(d, nptperyear, south = FALSE)
}
\arguments{
\item{d}{A data.table, should have \code{t} (compositing date) column
(\code{Date} variable).}

\item{nptperyear}{Integer, number of images per year}

\item{south}{Boolean. In south hemisphere, growing year is 1 July to the
following year 31 June; In north hemisphere, growing year is 1 Jan to 31 Dec.}
}
\value{
data.table
}
\description{
Add one year data in the head and tail
}
