% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seasonality.R
\name{check_seasonality}
\alias{check_seasonality}
\title{Check vegetation seasonality}
\usage{
check_seasonality(INPUT, IsPlot = F, pdat = INPUT, nf = 3, ...)
}
\description{
Check vegetation seasonality
}
