% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract_main.R
\name{curvefit}
\alias{curvefit}
\title{curve fit vegetation index (VI) time-series}
\usage{
curvefit(y, t = index(y), tout = t, meth = "BFGS", methods = c("spline",
  "beck", "elmore", "klos", "AG", "zhang"), ...)
}
\arguments{
\item{y}{Vegetation time-series index, numeric vector}

\item{t}{The corresponding doy of x}

\item{tout}{The output interpolated time.}
}
\description{
curve fit VI using 'spline', 'beck', 'elmore', 'klos', 'AG', 'Gu', 'zhang' methods
}
