% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvefits_extend.R
\name{curvefit}
\alias{curvefit}
\title{curve fit vegetation index (VI) time-series}
\usage{
curvefit(y, t = index(y), tout = t, meth = "BFGS",
  methods = c("spline", "beck", "elmore", "klos", "AG", "zhang"), ...)
}
\arguments{
\item{y}{Vegetation time-series index, numeric vector}

\item{t}{The corresponding doy of x}

\item{tout}{The output interpolated time.}

\item{meth}{optimization method}

\item{methods}{curve fitting method names, can be one or more of 'beck', 
'elmore', 'klos', 'AG', 'Gu', 'zhang'. Note that 'klos' has so many parameters.
Its optimization is not stable.}

\item{...}{other parameters passed to curve fitting function.}
}
\description{
curve fit VI using  methods
}
