% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvefits.R
\name{curvefits}
\alias{curvefits}
\title{Fine Curve fitting}
\usage{
curvefits(INPUT, brks, wFUN = wTSM, iters = 2, wmin = 0.2,
  nextent = 2, maxExtendMonth = 3, minExtendMonth = 1, minT = 0,
  methods = c("AG", "Beck", "Elmore", "Gu", "Klos", "Zhang"),
  qc = INPUT$w * 0 + 1, minPercValid = 0.2, print = TRUE, ...)
}
\arguments{
\item{INPUT}{A list object with the elements of 't', 'y', 'w', 'Tn' (option)
and 'ylu', returned by \code{check_input}.}

\item{brks}{A list object with the elements of 'fit' and 'dt', returned by
\code{season} or \code{season_3y}, which contains the growing season
dividing information.}

\item{wFUN}{weights updating function, can be one of 'wTSM', 'wChen' and
'wBisquare'.}

\item{iters}{How many times curve fitting is implemented.}

\item{wmin}{Double, minimum weigth (i.e. weight of snow, ice and cloud).}

\item{nextent}{Extend curve fitting window, until \code{nextent} good or
marginal element are found in previous and subsequent growing season.}

\item{maxExtendMonth}{Search good or marginal good values in previous and
subsequent `maxExtendMonth` period.}

\item{minExtendMonth}{Extending perid defined by \code{nextent} and
\code{maxExtendMonth} should be no shorter than \code{minExtendMonth}.
When all points of the input time-series are good value, then the extending
period will be too short. In that situation, we can't make sure the connection
between different growing seasons is smoothing.}

\item{minT}{Double, use night temperature Tn to define backgroud value.
Tn < minT is treated as ungrowing season.}

\item{methods}{Fine curve fitting methods, can be one or more of \code{c('AG', 
'Beck', 'Elmore', 'Gu', 'Klos', 'Zhang')}.}

\item{qc}{Factor (optional), only suit for MOD13A1. SummaryQA code for PhenoExtracr.
Other dataset, just leave qc as the default.}

\item{minPercValid}{If valid percentage is less than \code{minPercValid}, the
fits are set to NA.}

\item{print}{Whether to print progress information?}

\item{...}{Other parameters will be ignore.}
}
\value{
fits Multiple phenofit object.
}
\description{
Fine Curve fitting for INPUT time-series.
}
\examples{
data("MOD13A1")

dt <- tidy_MOD13.gee(MOD13A1$dt)
st <- MOD13A1$st

sitename <- dt$site[1]
d     <- dt[site == sitename, ] # get the first site data
sp    <- st[site == sitename, ] # station point
# global parameter
IsPlot = T
print  = F
nptperyear = 23
ypeak_min  = 0.05

dnew     <- add_HeadTail(d, nptperyear = 23) # add one year in head and tail
INPUT    <- check_input(dnew$t, dnew$y, dnew$w, nptperyear, 
                        maxgap = nptperyear/4, alpha = 0.02, wmin = 0.2)
}
