% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit.R
\name{D1}
\alias{D1}
\alias{D2}
\title{DD}
\usage{
D1(fit, ...)

D2(fit, ...)
}
\arguments{
\item{fit}{Curve fitting object returned by \code{curvefit}.}

\item{...}{ignored.}
}
\description{
D1 first order derivative, D2 second order derivative.
}
