% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative.R
\name{D}
\alias{D}
\alias{D1}
\alias{D2}
\alias{D1.phenofit}
\alias{D2.phenofit}
\title{D}
\usage{
D1(fit, analytical = TRUE, smoothed.spline = FALSE, ...)

D2(fit, analytical = TRUE, smoothed.spline = FALSE, ...)

\method{D1}{phenofit}(fit, analytical = TRUE, smoothed.spline = FALSE,
  ...)

\method{D2}{phenofit}(fit, analytical = TRUE, smoothed.spline = FALSE,
  ...)
}
\arguments{
\item{fit}{A curve fitting object returned by \code{curvefit}.}

\item{analytical}{If true, \code{numDeriv} package \code{grad} and \code{hess}
will be used; if false, \code{D1} and \code{D2} will be used.}

\item{smoothed.spline}{Whether apply \code{smooth.spline} first?}

\item{...}{Other parameters will be ignored.}
}
\description{
Get derivative of \code{phenofit} object.
\code{D1} first order derivative, \code{D2} second order derivative.
}
\details{
If \code{fit$fun} has no gradient function or \code{smoothed.spline = TRUE}, 
time-series smoothed by spline first, and get derivatives at last. 
If \code{fit$fun} exists and \code{analytical = TRUE}, \code{smoothed.spline} 
will be ignored.
}
