% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics.R
\name{f_goal}
\alias{f_goal}
\title{Common goal function of those curve fitting methods}
\usage{
f_goal(par, x, t, fun = c(doubleLog.elmore, doubleLog.beck, doubleLog.klos,
  doubleLog.gu, doubleLog.zhang, doubleAG), w, ylu, ...)
}
\description{
Common goal function of those curve fitting methods
}
