% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_goal.R
\name{f_goal}
\alias{f_goal}
\alias{f_goal_r}
\title{Goal function of fine curve fitting methods}
\usage{
f_goal(par, y, t, fun, w = NULL, ylu = NULL, ...)

f_goal_r(par, y, t, fun, w, ylu, ...)
}
\arguments{
\item{par}{A vector of parameters}

\item{y}{Numeric vector, vegetation index time-series}

\item{t}{Numeric vector, \code{Date} variable}

\item{fun}{A curve fitting function, can be one of \code{doubleAG},
\code{doubleLog.Beck}, \code{doubleLog.Elmore}, \code{doubleLog.Gu},
\code{doubleLog.Klos}, \code{doubleLog.Zhang}, see \code{\link[=Logistic]{Logistic()}}
for details.}

\item{w}{(optional) Numeric vector, weights of \code{y}. If not specified,
weights of all \code{NA} values will be \code{wmin}, the others will be 1.0.}

\item{ylu}{\code{ymin, ymax}, which is used to force \code{ypred} in the
range of \code{ylu}.}

\item{...}{others will be ignored.}
}
\value{
RMSE Root Mean Square Error of curve fitting values.
}
\description{
Goal function of fine curve fitting methods
}
\examples{
library(phenofit)
# simulate vegetation time-series
fFUN = doubleLog.Beck
par  = c(
    mn  = 0.1,
    mx  = 0.7,
    sos = 50,
    rsp = 0.1,
    eos = 250,
    rau = 0.1)
t    <- seq(1, 365, 8)
tout <- seq(1, 365, 1)
y <- fFUN(par, t)

par0 <- c(
    mn  = 0.15,
    mx  = 0.65,
    sos = 100,
    rsp = 0.12,
    eos = 200,
    rau = 0.12)
f_goal(par0, y, t, fFUN)
}
