% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics_fitting.R
\name{FitDL}
\alias{FitDL}
\alias{FitDL.Zhang}
\alias{FitAG}
\alias{FitDL.Beck}
\alias{FitDL.Elmore}
\alias{FitDL.Gu}
\alias{FitDL.Klos}
\title{Fine curve fitting functions.}
\usage{
FitDL.Zhang(y, t = index(y), tout = t, optimFUN = I_optimx,
  method = "nlm", w, ...)

FitAG(y, t = index(y), tout = t, optimFUN = I_optimx,
  method = "nlminb", w, ...)

FitDL.Beck(y, t = index(y), tout = t, optimFUN = I_optimx,
  method = "nlminb", w, ...)

FitDL.Elmore(y, t = index(y), tout = t, optimFUN = I_optimx,
  method = "nlminb", w, ...)

FitDL.Gu(y, t = index(y), tout = t, optimFUN = I_optimx,
  method = "nlminb", w, ...)

FitDL.Klos(y, t = index(y), tout = t, optimFUN = I_optimx,
  method = "BFGS", w, ...)
}
\arguments{
\item{y}{input vegetation index time-series.}

\item{t}{the corresponding doy(day of year) of y.}

\item{tout}{the output curve fitting time-series time steps.}

\item{optimFUN}{optimization function to solve curve fitting functions'
parameters. It's should be `optimx_fun`, or `optim_p`.}

\item{method}{method passed to `optimx` or `optim` function.}

\item{w}{weights}

\item{...}{other paraters passed to optimFUN, such as weights.}
}
\value{
list(pred, par, fun)
}
\description{
Fine curve fitting functions, e.g. double logistics, Asymmetric 
Gaussian. They are used to fit vegetation time-series in every growing season.
}
\references{
[1]. Beck, P.S.A., Atzberger, C., Hogda, K.A., Johansen, B., Skidmore, A.K.,
     2006. Improved monitoring of vegetation dynamics at very high latitudes:
     A new method using MODIS NDVI. Remote Sens. Environ.
     https://doi.org/10.1016/j.rse.2005.10.021.

[2]. Elmore, A.J., Guinn, S.M., Minsley, B.J., Richardson, A.D., 2012.
     Landscape controls on the timing of spring, autumn, and growing season
     length in mid-Atlantic forests. Glob. Chang. Biol. 18, 656-674.
     https://doi.org/10.1111/j.1365-2486.2011.02521.x. \cr

[3]. Gu, L., Post, W.M., Baldocchi, D.D., Black, T.A., Suyker, A.E., Verma,
     S.B., Vesala, T., Wofsy, S.C., 2009. Characterizing the Seasonal Dynamics
     of Plant Community Photosynthesis Across a Range of Vegetation Types,
     in: Noormets, A. (Ed.), Phenology of Ecosystem Processes: Applications
     in Global Change Research. Springer New York, New York, NY, pp. 35-58.
     https://doi.org/10.1007/978-1-4419-0026-5_2. \cr

[4]. https://github.com/kongdd/phenopix/blob/master/R/FitDoubleLogGu.R
}
