% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitting.R
\name{get_fitting}
\alias{get_fitting}
\alias{get_fitting.fFITs}
\title{getFittings}
\usage{
get_fitting(fit)

get_fitting.fFITs(fFITs)
}
\arguments{
\item{fit}{Object returned by \code{curvefits}.}

\item{fFITs}{\code{fFITs} object returned by \code{\link[=curvefit]{curvefit()}}.}
}
\description{
Get curve fitting data.frame
}
\examples{
library(phenofit)
# simulate vegetation time-series
fFUN = doubleLog.Beck
par  = c( mn  = 0.1, mx  = 0.7, sos = 50, rsp = 0.1, eos = 250, rau = 0.1)
t    <- seq(1, 365, 8)
tout <- seq(1, 365, 1)
y <- fFUN(par, t)
methods <- c("AG", "Beck", "Elmore", "Gu", "Zhang") # "Klos" too slow
fFITs <- curvefit(y, t, tout, methods)
# multiple years
fits  <- list(`2001` = fFITs, `2002` = fFITs)

l_param   <- get_param(fits)
d_GOF     <- get_GOF(fits)
d_fitting <- get_fitting(fits)
l_pheno   <- get_pheno(fits, "AG", IsPlot=TRUE)
}
