% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_main.R
\name{getparam}
\alias{getparam}
\title{Get parameters from curve fitting result}
\usage{
getparam(fit)
}
\arguments{
\item{fit}{Object returned by curvefits}
}
\description{
Get parameters from curve fitting result
}
