% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{GOF}
\alias{GOF}
\title{GOF}
\usage{
GOF(Y_obs, Y_sim, w, include.cv = FALSE)
}
\arguments{
\item{Y_obs}{Numeric vector, observations}

\item{Y_sim}{Numeric vector, corresponding simulated values}

\item{w}{Numeric vector, weights of every points}
}
\description{
Good of fitting
}
