% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimDL.R
\name{I_optimx}
\alias{I_optimx}
\title{I_optimx}
\usage{
I_optimx(prior, FUN, x, t, tout, method, debug = FALSE, ...)
}
\arguments{
\item{method}{Method passed to optimx function, which can be one of:
'BFGS','CG','Nelder-Mead','L-BFGS-B','nlm','nlminb', spg','ucminf',
'Rcgmin','Rvmmin','newuoa','bobyqa','nmkb','hjkb'}

\item{debug}{If debug is true, all.methods is also true for optimx}
}
\value{
result same as what optimx function returned
}
\description{
Interface of optimx package which unified optimization functions.
Should be caution that optimx speed is not so satisfied. So `I_optim` is also
present.
}
