% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifelse2.R
\name{ifelse2}
\alias{ifelse2}
\title{ifelse2}
\usage{
ifelse2(test, yes, no)
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}

\item{yes}{return values for true elements of test.}

\item{no}{return values for false elements of test.}
}
\description{
ternary operator just like java `test ? yes : no`
}
\examples{
x <- ifelse2(TRUE, 1:4, 1:10)

}
