% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_lambda.R
\name{init_lambda}
\alias{init_lambda}
\title{Initial lambda value of Whittaker smoother}
\usage{
init_lambda(y)
}
\arguments{
\item{y}{Numeric vector}
}
\description{
This function is only suitable for 16-day EVI time-series.
}
\examples{
library(phenofit)
data("MOD13A1")

dt <- tidy_MOD13.gee(MOD13A1$dt)
st <- MOD13A1$st

sitename <- dt$site[1]
d      <- dt[site == sitename, ] # get the first site data
lambda <- init_lambda(d$y)
}
\keyword{internal}
