% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics_fitting.R
\name{init_param}
\alias{init_param}
\title{init_param
Initialize double logistic function parameters}
\usage{
init_param(y, t, w)
}
\arguments{
\item{y}{input vegetation index time-series.}

\item{t}{the corresponding doy(day of year) of y.}

\item{w}{weights}
}
\description{
init_param
Initialize double logistic function parameters
}
