% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_wWHIT_lambda.R
\name{lambda_vcurve}
\alias{lambda_vcurve}
\alias{lambda_vcurve_jl}
\alias{lambda_cv_jl}
\title{lambda_vcurve}
\usage{
lambda_vcurve(
  y,
  w,
  d = 2,
  lg_lambda_min = 0.1,
  lg_lambda_max = 3,
  IsPlot = FALSE
)

lambda_vcurve_jl(y, w, lg_lambda_min = 0.1, lg_lambda_max = 3)

lambda_cv_jl(y, w, d = 2, lg_lambda_min = 0.1, lg_lambda_max = 3)
}
\description{
lambda_vcurve
}
\keyword{internal}
