% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{melt_list}
\alias{melt_list}
\alias{listk}
\alias{list.cbind}
\alias{list.rbind}
\title{melt_list}
\usage{
melt_list(list, var.name, na.rm = TRUE, ...)

listk(...)

list.cbind(x)

list.rbind(x)
}
\description{
melt_list

listk
}
