% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleLogistics_cpp.R
\name{logistic}
\alias{logistic}
\alias{doubleLog_Zhang}
\alias{doubleLog_AG}
\alias{doubleLog_Beck}
\alias{doubleLog_Elmore}
\alias{doubleLog_Gu}
\alias{doubleLog_Klos}
\title{Double logistics in Rcpp}
\usage{
logistic(par, t, pred)

doubleLog_Zhang(par, t, pred)

doubleLog_AG(par, t, pred)

doubleLog_Beck(par, t, pred)

doubleLog_Elmore(par, t, pred)

doubleLog_Gu(par, t, pred)

doubleLog_Klos(par, t, pred)
}
\arguments{
\item{par}{A vector of parameters}

\item{t}{A \code{Date} or numeric vector}

\item{pred}{Numeric Vector, predicted values}
}
\description{
Double logistics in Rcpp
}
\seealso{
\code{\link[=doubleLog.Beck]{doubleLog.Beck()}}
}
\keyword{internal}
