% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{melt_list}
\alias{melt_list}
\title{melt_list}
\usage{
melt_list(list, var.name, na.rm = TRUE, ...)
}
\arguments{
\item{list}{Data set to melt}

\item{var.name}{list names convert into var.name}

\item{na.rm}{Should NA values be removed from the data set? 
This will convert explicit missings to implicit missings.}

\item{...}{other parameters to melt.}
}
\description{
melt_list
}
