% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{merge_pdf}
\alias{merge_pdf}
\title{merge_pdf}
\usage{
merge_pdf(outfile = "RPlot.pdf", indir = "Figure", pattern = "*.pdf",
  del = FALSE)
}
\arguments{
\item{outfile}{String}

\item{indir}{Directory to search pdfs}

\item{pattern}{string used to match pdf filename}

\item{del}{Booolean. If true, after merge, original pdf files will be delete.}
}
\description{
rely on python pdfmerge package, \code{pip install pdfmerge}
}
