% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimDL.R
\name{optim_pheno}
\alias{optim_pheno}
\title{optim_pheno}
\usage{
optim_pheno(prior, FUN, x, t, tout, optimFUN = I_optim, method, w, ylu,
  iters = 2, wFUN = wTSM, nptperyear = 46, debug = FALSE, ...)
}
\value{
list(pred, par, fun)
}
\description{
Interface of optimization functions for double logistics and many other
curve fitting functions
}
\details{
If using weights updating method, need to set `nptperyear` as global parameter.
}
