% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract.R
\name{PhenoExtract}
\alias{PhenoExtract}
\title{PhenoExtract}
\usage{
PhenoExtract(fits, TRS = c(0.1, 0.2, 0.5, 0.6), analytical = TRUE,
  smoothed.spline = FALSE, IsPlot = FALSE, ...)
}
\arguments{
\item{fits}{Multiple \code{phenofit} object.}

\item{TRS}{Threshold for \code{PhenoTrs}.}

\item{analytical}{If true, \code{numDeriv} package \code{grad} and \code{hess}
will be used; if false, \code{D1} and \code{D2} will be used.}

\item{smoothed.spline}{Whether apply \code{smooth.spline} first?}

\item{IsPlot}{Boolean. Whether to plot figure?}

\item{...}{ignored.}
}
\value{
List of every year phenology metrics
}
\description{
Get yearly vegetation phenological metrics of a curve fitting method
}
