% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract.R
\name{PhenoTrs}
\alias{PhenoTrs}
\alias{PhenoDeriv}
\alias{PhenoGu}
\alias{PhenoKl}
\title{Phenology extraction functions}
\usage{
PhenoTrs(fit, approach = c("White", "Trs"), trs = 0.5, IsPlot = TRUE,
  ...)

PhenoDeriv(fit, analytical = TRUE, smoothed.spline = FALSE,
  IsPlot = TRUE, show.lgd = T, ...)

PhenoGu(fit, analytical = TRUE, smoothed.spline = FALSE,
  IsPlot = TRUE, ...)

PhenoKl(fit, analytical = TRUE, smoothed.spline = FALSE,
  IsPlot = TRUE, show.lgd = T, ...)
}
\arguments{
\item{fit}{A curve fitting object returned by \code{curvefit}.}

\item{approach}{to be used to calculate phenology metrics.
'White' (White et al. 1997) or 'Trs' for simple threshold.}

\item{trs}{threshold to be used for approach "Trs", in (0, 1).}

\item{IsPlot}{whether to plot?}

\item{...}{other parameters to PhenoPlot}

\item{analytical}{If true, \code{numDeriv} package \code{grad} and \code{hess}
will be used; if false, \code{D1} and \code{D2} will be used.}

\item{smoothed.spline}{Whether apply \code{smooth.spline} first?}

\item{show.lgd}{whether show figure lelend?}
}
\description{
Phenology extraction functions

PhenoDeriv

PhenoGu

PhenoKl
}
