% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract.R
\name{PhenoTrs}
\alias{PhenoTrs}
\alias{PhenoDeriv}
\alias{PhenoGu}
\alias{PhenoKl}
\title{Phenology extraction functions}
\usage{
PhenoTrs(fit, approach = c("White", "Trs"), trs = 0.5, IsPlot = TRUE,
  IsSmoothed = T, ...)

PhenoDeriv(fit, IsPlot = TRUE, smspline = TRUE, show.lgd = T, ...)

PhenoGu(fit, IsPlot = TRUE, smspline = TRUE, ...)

PhenoKl(fit, IsPlot = TRUE, show.lgd = T, ...)
}
\arguments{
\item{fit}{Curve fitting result returned by \code{curvefit}.}

\item{approach}{to be used to calculate phenology metrics. 
'White' (White et al. 1997) or 'Trs' for simple threshold.}

\item{trs}{threshold to be used for approach "Trs", in (0, 1).}

\item{IsPlot}{whether to plot?}

\item{IsSmoothed}{Boolean. If false, positive derivative in spring and negative 
derivative will be not applied.}

\item{...}{other parameters to PhenoPlot}

\item{smspline}{If smspline = TRUE or attr(fit$fun, 'gradient') == null,
it will use smooth.spline to get der1 and der2.}

\item{show.lgd}{whether show figure lelend?}
}
\description{
Phenology extraction functions

PhenoDeriv

PhenoGu

PhenoKl
}
