% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit-package.R
\docType{package}
\name{phenofit}
\alias{phenofit}
\alias{phenofit-package}
\title{phenofit}
\description{
Vegetation phenology package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eco-hydro/phenofit}
  \item Report bugs at \url{https://github.com/eco-hydro/phenofit/issues}
}

}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

Authors:
\itemize{
  \item Mingzhong Xiao \email{xmingzh@mail2.sysu.edu.cn}
  \item Yongqiang Zhang \email{yongqiang.zhang2014@gmail.com}
  \item Xihui Gu \email{guxh@cug.edu.cn}
  \item Jianjian Cui \email{cuijj6@mail2.sysu.edu.cn}
}

}
\keyword{Vegetation}
\keyword{internal}
\keyword{package}
\keyword{phenology}
