% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_process.R
\name{phenofit_plot}
\alias{phenofit_plot}
\title{phenofit_plot}
\usage{
phenofit_plot(obj, type = "all", methods, title = NULL,
  title.ylab = "Vegetation Index", Isplot = TRUE, show.legend = TRUE,
  newpage = TRUE)
}
\arguments{
\item{obj}{\code{fFIT}}

\item{type}{one of c("season", "fitting", "pheno", "all")}

\item{title}{String, title of figure.}

\item{show.legend}{If now show legend, ggplot object will be returned, else
grid object will be returned.}
}
\description{
phenofit_plot
}
