% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_plot.R
\name{phenofit_plot}
\alias{phenofit_plot}
\title{phenofit_plot}
\usage{
phenofit_plot(obj, type = "all", methods, title = NULL,
  title.ylab = "Vegetation Index", IsPlot = TRUE, show.legend = TRUE,
  newpage = TRUE)
}
\arguments{
\item{obj}{\code{fFIT}}

\item{type}{one of c("season", "fitting", "pheno", "all")}

\item{methods}{Fine curve fitting methods, can be one or more of
\code{c('AG', 'Beck', 'Elmore', 'Gu', 'Klos', 'Zhang')}.}

\item{title}{String, title of figure.}

\item{title.ylab}{String, title of \code{xlab} and \code{ylab}.}

\item{IsPlot}{boolean. If false, a ggplot object will be returned.}

\item{show.legend}{If now show legend, ggplot object will be returned, else
grid object will be returned.}

\item{newpage}{boolean, whether draw figure in a new page?}
}
\description{
phenofit_plot
}
