% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_process.R
\name{phenofit_process}
\alias{phenofit_process}
\alias{getsite_data}
\alias{getsite_INPUT}
\alias{phenofit_season}
\alias{phenofit_finefit}
\title{phenofit_process}
\usage{
getsite_data(df, sitename, dateRange = NULL)

getsite_INPUT(df, st, sitename, nptperyear, dateRange = NULL)

phenofit_season(INPUT, options, IsPlot = FALSE, ...)

phenofit_finefit(INPUT, brks, options, ...)

phenofit_process(options, dateRange = c(as.Date("2010-01-01"),
  as.Date("2014-12-31")), nsite = -1, progress = NULL, ...)
}
\arguments{
\item{dateRange}{Date vector, filter input in the \code{dateRange}}

\item{INPUT}{An object returned by \code{check_input}}

\item{options}{options of phenofit}

\item{IsPlot}{whether to plot season dividing procedure?}

\item{...}{other parameters to \code{\link{season_mov}} or \code{\link{season}}}

\item{brks}{object returned by \code{\link{season_mov}} and \code{\link{season}}}

\item{nsite}{the max number of sites to process. \code{-1} means all sites.}
}
\value{
same object returned by \code{\link{season_mov}} and \code{\link{season}}
}
\description{
\describe{
  \item{phenofit_season}{ Calculate growing season dividing information}
  \item{phenofit_finefit}{Fine fitting}
}
}
\examples{
\dontrun{
file_json <- system.file('shiny/phenofit/perference/phenofit_setting.json', package = "phenofit")
options <- setting.read(file_json)
r <- phenofit_process(options, nsite=2)
}
}
