% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_TS.avhrr.R
\name{phenofit_TS.avhrr}
\alias{phenofit_TS.avhrr}
\title{phenofit_TSM.avhrr}
\usage{
phenofit_TS.avhrr(options, dateRange = c(as.Date("2010-01-01"),
  as.Date("2014-12-31")), I_part = NULL, outdir = ".",
  exportType = "all", overwrite = FALSE, .progress = NULL,
  .parallel = FALSE, ...)
}
\arguments{
\item{options}{options of phenofit needed to export.}

\item{dateRange}{Date vector, \code{[date_begin, date_end]}. filter input in the
range of \code{dateRange}}

\item{I_part}{numeric vector, index of pixels needed to run.}

\item{outdir}{directory of output}

\item{exportType}{could be one of \code{"all", "pheno"}. If "all" used, all
also will be exported. Note that exported fitting series is daily scale, which
is quite large.}

\item{overwrite}{boolean, whether overwrite previous output?}

\item{.progress}{boolean}

\item{.parallel}{boolean}

\item{...}{ignored}
}
\description{
phenofit_TSM.avhrr
}
