% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenofit.R
\name{plot_phenofit}
\alias{plot_phenofit}
\title{plot_phenofit}
\usage{
plot_phenofit(
  obj,
  type = "all",
  methods,
  title = NULL,
  ylab = "Vegetation Index",
  IsPlot = TRUE,
  show.legend = TRUE,
  newpage = TRUE,
  ...
)
}
\arguments{
\item{obj}{\code{phenofit} object, list(INPUT, fit, seasons)}

\item{type}{one of c("season", "fitting", "pheno", "all")}

\item{methods}{fine fitting functions, e.g., \code{c("AG", "Zhang", "Beck", "Elmore", "Gu")}}

\item{title}{String, title of figure.}

\item{IsPlot}{boolean. If false, a ggplot object will be returned.}

\item{show.legend}{If now show legend, ggplot object will be returned, else
grid object will be returned.}

\item{newpage}{boolean, whether draw figure in a new page?}

\item{...}{other parameters to \code{\link[=plot_curvefits]{plot_curvefits()}}}
}
\description{
plot_phenofit
}
\keyword{internal}
