% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoExtract_main.R
\name{plot_phenofit}
\alias{plot_phenofit}
\title{plot_phenofit}
\usage{
plot_phenofit(fit, d, title = NULL, plotly = F)
}
\arguments{
\item{fit}{data from phenofit_site}
}
\description{
plot_phenofit
}
