% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenofit.R
\name{plot_phenofit}
\alias{plot_phenofit}
\title{plot_phenofit}
\usage{
plot_phenofit(fit, d, title = NULL, show.legend = T)
}
\arguments{
\item{fit}{Object returned by \code{curvefits}.}

\item{d}{original input data}

\item{title}{String}

\item{show.legend}{Boolean}
}
\description{
plot_phenofit
}
