% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{qc_levels}
\alias{qc_levels}
\alias{qc_colors}
\alias{qc_shapes}
\title{qc level, color and shape}
\format{
An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{numeric} of length 6.
}
\usage{
qc_levels

qc_colors

qc_shapes
}
\description{
qc level, color and shape
}
\keyword{datasets}
\keyword{internal}
