% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcFUN.R
\name{getBits}
\alias{getBits}
\alias{qc_summary}
\alias{qc_5l}
\alias{qc_NDVIv4}
\title{Initial weights according to qc}
\usage{
getBits(x, start, end = start)

qc_summary(QA, wmin = 0.1)

qc_5l(QA, wmin = 0.1)

qc_NDVIv4(QA)
}
\arguments{
\item{x}{Binary value}

\item{start}{Bit starting position}

\item{end}{Bit ending position}

\item{QA}{quality control variable}
}
\value{
weigths
}
\description{
\describe{
  \item{getBits}{Extract bitcoded QA information from bin value}
  \item{qc_summary}{Initial weigths based on Quality reliability of VI pixel, 
suit for MOD13A1, MOD13A2 and MOD13Q1.}
  \item{qc_5l}{Initial weights based on Quality control of five-level 
confidence score, suit for MCD15A3H(LAI) and MOD17A2H(GPP)}
  \item{qc_NDVIv4}{For NDVIv4}
}
}
