% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcFUN.R
\name{getBits}
\alias{getBits}
\alias{qc_summary}
\alias{qc_StateQA}
\alias{qc_5l}
\alias{qc_NDVIv4}
\title{Initial weights according to qc}
\usage{
getBits(x, start, end = start)

qc_summary(QA, wmin = 0.2)

qc_StateQA(QA, wmin = 0.2)

qc_5l(QA, wmin = 0.2)

qc_NDVIv4(QA)
}
\arguments{
\item{x}{Binary value}

\item{start}{Bit starting position, count from zero}

\item{end}{Bit ending position}

\item{QA}{quality control variable}

\item{wmin}{Double, minimum weigth (i.e. weight of snow, ice and cloud).}
}
\value{
A list object with
\item{weigths}{Double vector, initial weights}
\item{QC_flag}{Factor vector, with the level of \code{c("snow", "cloud", "shadow", "aerosol", "marginal", "good")}}
}
\description{
\describe{
  \item{getBits}{Extract bitcoded QA information from bin value}
  \item{qc_summary}{Initial weigths based on Quality reliability of VI pixel, 
suit for MOD13A1, MOD13A2 and MOD13Q1 (SummaryQA band).}
  \item{qc_5l}{Initial weights based on Quality control of five-level 
confidence score, suit for MCD15A3H(LAI, FparLai_QC), MOD17A2H(GPP, Psn_QC) 
and MOD16A2(ET, ET_QC).}
  \item{qc_NDVIv4}{For NDVIv4}
  \item{qc_StateQA}{Initial weights based on `StateQA`, suit for MOD09A1, MYD09A1. }
}
}
