% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{R2_sign}
\alias{R2_sign}
\title{Critical value of determined correlation}
\usage{
R2_sign(n, NumberOfPredictor = 2, alpha = 0.05)
}
\arguments{
\item{n}{length of observation.}

\item{NumberOfPredictor}{Number of predictor, including constant.}

\item{alpha}{significant level.}
}
\value{
\code{F} statistic and \code{R2} at significant level.
}
\description{
Critical value of determined correlation
}
\examples{
R2_critical <- R2_sign(30, NumberOfPredictor = 2, alpha = 0.05)
}
\references{
Chen Yanguang (2012), Geographical Data analysis with MATLAB.
}
\keyword{internal}
