% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\title{season}
\usage{
season(INPUT, nptperyear = 46, south = FALSE, FUN = whitsmw2,
  wFUN = wTSM, iters = 2, wmin = 0.1, lambda, nf = 2,
  frame = floor(nptperyear/7) * 2 + 1, IsPlot = FALSE, plotdat = INPUT,
  print = FALSE, minpeakdistance = nptperyear/6, ymax_min = 0.5,
  rymin_less = 0.6, threshold_max = 0.2, threshold_min = 0.05,
  max_MaxPeaksperyear = 2, max_MinPeaksperyear = 3, ...)
}
\arguments{
\item{INPUT}{returned list(t, y, w, ylu) object from `check_input`}

\item{FUN}{Curve fitting functions, call be one of `sgfitw`, `whitsmw2` and `wHANTS`.}

\item{plotdat}{Is IsPlot = true, plotdata need to be passed. 
Since, y, w have been changed.}

\item{minpeakdistance}{The minimum distance (in indices) peaks have to have
to be counted. If the distance of two maximum extreme value less than
`minpeakdistance`, only the real maximum value will be left.}

\item{rymin_less}{y_tough < rymin_less * A}

\item{threshold_max}{Similar as `threshold_min`, The maximum threshold should
be greater than `threshold_max`.}

\item{threshold_min}{Threshold is defined as the difference of peak value with
trough value. There are two threshold (left and right). The minimum threshold
should be greater than threshold_min.}

\item{...}{Other parameters passed to findpeaks, e.g. 'threshold',
'minpeakheight', 'minpeakdistance'}
}
\value{
list(dt, di)
$whit
}
\description{
First smooth VI timeseries by weighted whittaker, then use findpeak function
to get the local maximum and local minimum values. Two local minimum defined
a growing season. If two local minimum(maximum) was adjacent, then only the
smaller(biger) is left.
}
\details{
Then according to season pos, based to local maximum position divide yearly
growing season. lambda need to set carefully.

divide growing season circle just like TIMESAT season function
}
