% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoBrks.R
\name{season}
\alias{season}
\title{season}
\usage{
season(INPUT, lambda, nptperyear = 46, south = FALSE, iters = 2,
  wFUN = wTSM, IsPlot = TRUE, minpeakdistance = nptperyear/6,
  ymax_min = 0.5, rymin_less = 0.6, threshold_max = 0.2,
  threshold_min = 0.05, max_MaxPeaksperyear = 2, max_MinPeaksperyear = 3,
  plotdat = INPUT, ...)
}
\arguments{
\item{INPUT}{returned list(t, y, w, ylu) object from `check_input`}

\item{rymin_less}{y_tough < rymin_less * A}

\item{plotdat}{Is IsPlot = true, plotdata need to be passed. 
Since, y, w have been changed.}

\item{...}{Other parameters passed to findpeaks, e.g. 'threshold',
'minpeakheight', 'minpeakdistance'}

\item{TRS}{multiply A (diff(ylu)) becomes the threshold parameter of
`findpeak` function}
}
\value{
list(dt, di)
$whit
}
\description{
First smooth VI timeseries by weighted whittaker, then use findpeak function
to get the local maximum and local minimum values. Two local minimum defined
a growing season. If two local minimum(maximum) was adjacent, then only the
smaller(biger) is left.
}
\details{
Then according to season pos, based to local maximum position divide yearly
growing season. lambda need to set carefully.

divide growing season circle just like TIMESAT season function
}
