% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_options}
\alias{set_options}
\alias{get_options}
\title{set and get phenofit option}
\usage{
set_options(...)

get_options(names = NULL)
}
\arguments{
\item{options}{\itemize{
\item style: font style, one of "EN" and "CH"
\item family: font family name, e.g. "Times"
\item family_CH: Chinese fontname for panel.barchartFreq yaxis title
}}
}
\description{
set and get phenofit option
}
\examples{
set_options(verbose_curvefit = FALSE)
get_options("verbose_season")
}
