% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_options}
\alias{set_options}
\alias{get_options}
\title{set and get phenofit option}
\usage{
set_options(...)

get_options(names = NULL)
}
\arguments{
\item{...}{list of phenofit options
FUN_season: character, \code{season_mov} or \code{season}
rFUN: character, rough fitting function. \code{smooth_wWHIT}, \code{smooth_wSG} or \code{smooth_wHANTs}.}

\item{names}{vector of character, names of options}
}
\description{
set and get phenofit option
}
\examples{
set_options(verbose_curvefit = FALSE)
get_options("verbose_season")
}
