% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_whit_lambda.R
\name{v_curve}
\alias{v_curve}
\title{update 20180605 add weights updating to whittaker lambda selecting}
\usage{
v_curve(INPUT, nptperyear, llas, d = 2, show = F, wFUN = wTSM,
  iters = 2)
}
\description{
update 20180605 add weights updating to whittaker lambda selecting
}
