% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightUpdate.R
\name{wBECK}
\alias{wBECK}
\title{Beck et al., (2006) weigths updating method}
\usage{
wBECK(y, yfit, w, ...)
}
\description{
Beck et al., (2006) weigths updating method
}
\references{
[1]. Beck, P.S.A., Atzberger, C., Høgda, K.A., Johansen, B., Skidmore, A.K., 
     2006. Improved monitoring of vegetation dynamics at very high latitudes: 
     A new method using MODIS NDVI. Remote Sens. Environ. 
     https://doi.org/10.1016/j.rse.2005.10.021 \cr
[2]. https://github.com/kongdd/phenopix/blob/master/R/FitDoubleLogBeck.R
}
